requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-login
Flask-testing
Flask-Assets
Flask-admin
Flask-Migrate
Flask-Security

psycopg2
py-bcrypt

fabric

jsmin
cssmin
pyscss